+++ tests/testthat/test-col-spec.R
@@ -222,7 +222,7 @@ test_that("long expressions are wrapped (597)", {
   expect_equal(format(cols(a = col_factor(levels = c("apple", "pear", "banana", "peach", "apricot", "orange", "plum"), ordered = TRUE))),
 'cols(
   a = col_factor(levels = c("apple", "pear", "banana", "peach", "apricot", "orange", "plum"
-    ), ordered = TRUE)
+    ), ordered = TRUE, include_na = FALSE)
 )
 ')
 })