+++ tests/testthat/test-write-lines.R
@@ -93,3 +93,27 @@ test_that("write_file with raw round trips with an empty vector", {
 
   expect_equal(read_file_raw(tmp), x)
 })
+
+test_that("write_lines can write to compressed files", {
+
+  mt <- read_lines(readr_example("mtcars.csv.bz2"))
+
+  filename <- file.path(tempdir(), "mtcars.csv.bz2")
+  on.exit(unlink(filename))
+  write_lines(mt, filename)
+
+  expect_true(is_bz2_file(filename))
+  expect_equal(mt, read_lines(filename))
+})
+
+test_that("write_file can write to compressed files", {
+
+  mt <- read_file(readr_example("mtcars.csv.bz2"))
+
+  filename <- file.path(tempdir(), "mtcars.csv.bz2")
+  on.exit(unlink(filename))
+  write_file(mt, filename)
+
+  expect_true(is_bz2_file(filename))
+  expect_equal(mt, read_file(filename))
+})